created rag service + remove quivr_core from packages
