+++ backend/api/quivr_api/modules/rag_service/__init__.py
@@ -1,0 +1,3 @@
+from .rag_service import RAGService
+
+__all__ = ["RAGService"]
